+++ lib/view/book_screen/elected_member_profile/elected_member_profile.dart
@@ -3,6 +3,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:get/get.dart';
 import 'package:rtd_project/core/color/colors.dart';
+import 'package:rtd_project/core/common_widget/appbar.dart';
 import 'package:rtd_project/core/constraints/conatrints.dart';
 import 'package:url_launcher/url_launcher.dart';
 
@@ -18,7 +19,16 @@ class BoardMemberDetailsScreen extends StatelessWidget {
       backgroundColor: baseColor,
       body: Column(
         children: [
-          appbar(context),
+          CustomAppBar(
+              leading: IconButton(
+                onPressed: () => Get.back(),
+                icon: const Icon(
+                  Icons.arrow_back,
+                  color: whiteColor,
+                  size: 30,
+                ),
+              ),
+              title: 'Member Book'),
           Expanded(
             child: Container(
               decoration: const BoxDecoration(
