+++ lib/view/notification/notification.dart
@@ -5,10 +5,14 @@ import 'package:lottie/lottie.dart';
 import 'package:rtd_project/controller/loan/loan_request_detail_controller.dart';
 import 'package:rtd_project/controller/notification/notification_controller.dart';
 import 'package:rtd_project/core/color/colors.dart';
+import 'package:rtd_project/core/common_widget/appbar.dart';
 import 'package:rtd_project/core/constraints/conatrints.dart';
 import 'package:rtd_project/util/theme.dart';
+import 'package:rtd_project/util/toast.dart';
 import 'package:rtd_project/view/notification/widgets/notification_widget.dart';
 
+import '../../backend/model/notification_model/notification_model.dart';
+import '../../controller/loan/profile_loan_data_controller.dart';
 import '../../controller/notification/surety_view_controller.dart';
 import '../../helper/router.dart';
 import 'notification_details_screen/notification_details.dart';
@@ -21,7 +25,6 @@ class NotificationScreen extends StatefulWidget {
 }
 
 class _NotificationScreenState extends State<NotificationScreen> {
-  bool clikable = false;
   @override
   Widget build(BuildContext context) {
     return SafeArea(
@@ -29,10 +32,57 @@ class _NotificationScreenState extends State<NotificationScreen> {
         backgroundColor: baseColor,
         body: GetBuilder<NotificationController>(
           builder: (controller) {
+            List<NotificationDetails>? notificationList =
+                controller.notification?.data;
             return controller.loading != true
                 ? Column(
                     children: [
-                      appbar(context, controller),
+                      CustomAppBar(
+                        leading: IconButton(
+                          onPressed: Get.back,
+                          icon: const Icon(
+                            Icons.arrow_back,
+                            color: whiteColor,
+                            size: 30,
+                          ),
+                        ),
+                        title: 'Notification',
+                        trailing: Stack(
+                          alignment: AlignmentDirectional.center,
+                          children: [
+                            IconButton(
+                              onPressed: () {
+                                // Get.toNamed(AppRouter.getSuretyRequestDetailsRoute());
+                              },
+                              icon: const Icon(
+                                Icons.notifications_none,
+                                color: whiteColor,
+                                size: 35,
+                              ),
+                            ),
+                            controller.notificationCount != 0
+                                ? Positioned(
+                                    top: 3.h,
+                                    right: 4.w,
+                                    child: Container(
+                                      height: 20.h,
+                                      width: 20.w,
+                                      decoration: const BoxDecoration(
+                                          color: Colors.red,
+                                          shape: BoxShape.circle),
+                                      child: Center(
+                                          child: Text(
+                                        '${controller.notificationCount ?? 0}',
+                                        style: const TextStyle(
+                                            color: whiteColor,
+                                            fontWeight: FontWeight.bold),
+                                      )),
+                                    ),
+                                  )
+                                : const Offstage(),
+                          ],
+                        ),
+                      ),
                       Expanded(
                         child: Container(
                           padding: EdgeInsets.only(
@@ -44,24 +94,21 @@ class _NotificationScreenState extends State<NotificationScreen> {
                               topEnd: Radius.circular(50),
                             ),
                           ),
-                          child: controller.notification?.data?.length != 0
+                          child: notificationList?.length != 0
                               ? ListView.separated(
                                   padding: EdgeInsets.only(bottom: 10.h),
                                   shrinkWrap: true,
-                                  itemCount:
-                                      controller.notification!.data!.length,
+                                  itemCount: notificationList!.length,
                                   physics: const BouncingScrollPhysics(),
                                   separatorBuilder: (context, index) =>
                                       kSizedBoxH,
                                   itemBuilder: (context, index) {
-                                    final type = controller
-                                        .notification?.data?[index].type;
+                                    final type = notificationList[index].type;
                                     if (type == 4) {
                                       return containerWithClickable(
                                         onTap: () {
                                           controller.markNotificationSeen(
-                                            controller
-                                                .notification!.data![index].id!,
+                                            notificationList[index].id!,
                                           );
 
                                           Get.delete<SuretyViewController>(
@@ -70,16 +117,15 @@ class _NotificationScreenState extends State<NotificationScreen> {
                                               AppRouter
                                                   .getSuretyRequestDetailsRoute(),
                                               arguments: [
-                                                controller.notification!
-                                                    .data![index].details?.id,
-                                                controller
-                                                    .notification!.data![index],
+                                                notificationList[index]
+                                                    .details
+                                                    ?.id,
+                                                notificationList[index],
                                                 false
                                               ]);
                                         },
-                                        90.h,
-                                        controller
-                                            .notification?.data?[index].message,
+                                        110.h,
+                                        notificationList[index].message,
                                         clikable: 'View Details',
                                         index: index,
                                         controller: controller,
@@ -88,35 +134,35 @@ class _NotificationScreenState extends State<NotificationScreen> {
                                       return NotificationContainer(
                                         onTap: () {
                                           controller.markNotificationSeen(
-                                            controller
-                                                .notification!.data![index].id!,
+                                            notificationList[index].id!,
                                           );
+                                          Get.delete<ProfileLoanDataController>(
+                                              force: true);
                                           Get.toNamed(
                                               AppRouter
                                                   .getProfileLoanDetailsRoute(),
                                               arguments: [
-                                                controller.notification!
-                                                    .data![index].details!.id!
+                                                notificationList[index]
+                                                    .details!
+                                                    .id!
                                               ]);
                                         },
                                         height: 110.h,
-                                        title: controller
-                                            .notification?.data?[index].message,
+                                        title: notificationList[index].message,
                                         controller: controller,
                                         index: index,
                                         text2:
-                                            'Loan Amount :  ${controller.notification?.data?[index].details!.amount}',
+                                            'Loan Amount :  ${notificationList[index].details!.amount}',
                                         text3:
-                                            'Purpose : ${controller.notification?.data?[index].details!.purpose}',
+                                            'Purpose : ${notificationList[index].details!.purpose}',
                                         text4:
-                                            'Repayment Date :${controller.notification?.data?[index].details!.dueDate}',
+                                            'Repayment Date :${notificationList[index].details!.dueDate}',
                                       );
                                     } else if (type == 1) {
                                       return NotificationContainer(
                                         onTap: () {
                                           controller.markNotificationSeen(
-                                            controller
-                                                .notification!.data![index].id!,
+                                            notificationList[index].id!,
                                           );
                                           Get.delete<
                                                   LoanRequestDetailsController>(
@@ -125,38 +171,39 @@ class _NotificationScreenState extends State<NotificationScreen> {
                                               AppRouter
                                                   .getLoanRequestDetailsRoutes(),
                                               arguments: [
-                                                controller.notification!
-                                                    .data![index].details!.id,
+                                                notificationList[index]
+                                                    .details!
+                                                    .id,
                                                 true
                                               ]);
                                         },
                                         height: 110.h,
-                                        title: controller
-                                            .notification?.data?[index].message,
+                                        title: notificationList[index].message,
                                         controller: controller,
                                         index: index,
                                         text2:
-                                            'Loan Amount : ${controller.notification?.data?[index].details?.amount ?? "NA"}',
+                                            'Loan Amount : ${notificationList[index].details?.amount ?? "NA"}',
                                         text3:
-                                            'Purpose :  ${controller.notification?.data?[index].details?.purpose ?? "NA"}',
+                                            'Purpose :  ${notificationList[index].details?.purpose ?? "NA"}',
                                       );
                                     } else if (type == 3) {
-                                      final inputFormat = controller
-                                          .notification!.data?[index].date
-                                          .toString();
+                                      final inputFormat =
+                                          notificationList[index]
+                                              .date
+                                              .toString();
                                       return NotificationContainer(
                                         onTap: () {
                                           controller.markNotificationSeen(
-                                              controller.notification!
-                                                  .data![index].id!);
+                                              notificationList[index].id!);
                                           Get.delete<SuretyViewController>(
                                               force: true);
                                           Get.toNamed(
                                               AppRouter
                                                   .getNotificationPollRoute(),
                                               arguments: [
-                                                controller.notification!
-                                                    .data?[index].details!.id,
+                                                notificationList[index]
+                                                    .details!
+                                                    .id,
                                                 null,
                                                 true
                                               ]);
@@ -164,43 +211,42 @@ class _NotificationScreenState extends State<NotificationScreen> {
                                         height: 110.h,
                                         index: index,
                                         controller: controller,
-                                        title: controller
-                                            .notification?.data?[index].message,
+                                        title: notificationList[index].message,
                                         endDate: inputFormat,
                                         clickable: 'View Details',
                                       );
                                     } else if (type == 6) {
                                       return NotificationContainer(
-                                        onTap: () => controller
-                                            .markNotificationSeen(controller
-                                                .notification!
-                                                .data![index]
-                                                .id!),
+                                        onTap: () {
+                                          controller.markNotificationSeen(
+                                              notificationList[index].id!);
+                                          Get.delete<ProfileLoanDataController>(
+                                              force: true);
+                                          Get.toNamed(
+                                              AppRouter
+                                                  .getProfileLoanDetailsRoute(),
+                                              arguments: [
+                                                notificationList[index]
+                                                    .details!
+                                                    .id!
+                                              ]);
+                                        },
                                         height: 70.h,
-                                        title: controller
-                                            .notification?.data?[index].message,
+                                        title: notificationList[index].message,
                                         text1:
-                                            "Date : ${controller.notification?.data?[index].date}",
+                                            "Date : ${notificationList[index].date}",
                                         controller: controller,
                                         index: index,
                                       );
                                     }
                                     return containerWithClickable(onTap: () {
                                       controller.markNotificationSeen(
-                                        controller
-                                            .notification!.data![index].id!,
+                                        notificationList[index].id!,
                                       );
                                       Get.to(const NotificationDetailsScreen(),
-                                          arguments: [
-                                            controller
-                                                .notification!.data![index]
-                                          ]);
-                                    },
-                                        110.h,
-                                        controller
-                                            .notification?.data?[index].message,
-                                        controller: controller,
-                                        index: index);
+                                          arguments: [notificationList[index]]);
+                                    }, 110.h, notificationList[index].message,
+                                        controller: controller, index: index);
                                   },
                                 )
                               : Column(
@@ -432,11 +478,15 @@ class _NotificationScreenState extends State<NotificationScreen> {
     }
     return GestureDetector(
       onTap: () {
-        Get.delete<LoanRequestDetailsController>(force: true);
-        Get.toNamed(AppRouter.getLoanRequestDetailsRoutes(), arguments: [
-          controller.notification!.data![index].details!.id,
-          false
-        ]);
+        if (controller.notification!.data![index].details == null) {
+          showToast('Loan details not available from the server');
+        } else {
+          Get.delete<LoanRequestDetailsController>(force: true);
+          Get.toNamed(AppRouter.getLoanRequestDetailsRoutes(), arguments: [
+            controller.notification!.data![index].details!.id,
+            false
+          ]);
+        }
       },
       child: Container(
         height: height,
@@ -486,62 +536,6 @@ class _NotificationScreenState extends State<NotificationScreen> {
             )
           ],
         ),
-      ),
-    );
-  }
-
-  Container appbar(BuildContext context, NotificationController controller) {
-    return Container(
-      margin: EdgeInsets.only(top: 10.h),
-      child: Row(
-        mainAxisAlignment: MainAxisAlignment.spaceBetween,
-        children: [
-          IconButton(
-              onPressed: () => Get.back(),
-              icon: const Icon(
-                Icons.arrow_back,
-                color: whiteColor,
-                size: 30,
-              )),
-          Text(
-            'Notification',
-            style: Theme.of(context).textTheme.displaySmall!.copyWith(
-                color: Colors.white, fontWeight: FontWeight.bold, fontSize: 20),
-          ),
-          Stack(
-            alignment: AlignmentDirectional.center,
-            children: [
-              IconButton(
-                onPressed: () {
-                  // Get.toNamed(AppRouter.getSuretyRequestDetailsRoute());
-                },
-                icon: const Icon(
-                  Icons.notifications_none,
-                  color: whiteColor,
-                  size: 35,
-                ),
-              ),
-              controller.notificationCount != 0
-                  ? Positioned(
-                      top: 3.h,
-                      right: 4.w,
-                      child: Container(
-                        height: 20.h,
-                        width: 20.w,
-                        decoration: const BoxDecoration(
-                            color: Colors.red, shape: BoxShape.circle),
-                        child: Center(
-                            child: Text(
-                          '${controller.notificationCount ?? 0}',
-                          style: const TextStyle(
-                              color: whiteColor, fontWeight: FontWeight.bold),
-                        )),
-                      ),
-                    )
-                  : const Offstage(),
-            ],
-          ),
-        ],
       ),
     );
   }
